\name{potency}
\alias{potency}
\title{ Return or compute potency }
\description{ The function computes the potency as described in the Ph.Eur. }
\usage{ potency(object) }
\arguments{
  \item{object}{ Object of class \link{pla-class} or \link{plaFit-class} } }
\details{ See page 480 in Ph.Eur. }
\value{ A list of lists with variables named as at page 480 in Ph.Eur. }
\references{  Chapter 5.3. Statistical analysis.
  In EUROPEAN PHARMACOPOEIA 5.0, 2004; 475-504. }
\author{Jens Henrik Badsberg}
\keyword{ design }
